Errores porongas de Pinato arreglados

+++ src/GUI/bajaCotizacion.java
@@ -74,9 +74,16 @@ public class bajaCotizacion extends javax.swing.JFrame {
             Class[] types = new Class [] {
                 java.lang.Object.class, java.lang.String.class, java.lang.Double.class, java.lang.Object.class
             };
+            boolean[] canEdit = new boolean [] {
+                false, false, false, false
+            };
 
             public Class getColumnClass(int columnIndex) {
                 return types [columnIndex];
+            }
+
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
+                return canEdit [columnIndex];
             }
         });
         jScrollPane1.setViewportView(jTable1);
@@ -132,26 +139,29 @@ public class bajaCotizacion extends javax.swing.JFrame {
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         // TODO add your handling code here:
-        int input = javax.swing.JOptionPane.showConfirmDialog(null, "¿Desea continuar?", "Seleccione una opción",
-                javax.swing.JOptionPane.YES_NO_CANCEL_OPTION, javax.swing.JOptionPane.YES_NO_CANCEL_OPTION);
-
-        if (input == 0) {
-            DefaultTableModel mdl = (DefaultTableModel) jTable1.getModel();
-            Cotizacion c = (Cotizacion) jTable1.getValueAt(jTable1.getSelectedRow(), 3);
-            //Preguntar si la cotización no está en uso antes de borrar.
-            List<Factura> facturas = Conexion.getInstance().listarFacturasSinFechaADolares();
-            for (Factura f : facturas) {
-                if ((float) c.getImporte() == f.getCotizacion()) {
-                    javax.swing.JOptionPane.showMessageDialog(null, "No puede borrar una cotización si tiene una o más facturas asignadas.", "Error", javax.swing.JOptionPane.ERROR_MESSAGE);
-                    return;
+        if (jTable1.getSelectedRow() == -1) {
+            javax.swing.JOptionPane.showMessageDialog(null, "Debe seleccionar la cotización que desea eliminar.", "Error", javax.swing.JOptionPane.ERROR_MESSAGE);
+        } else {
+            int input = javax.swing.JOptionPane.showConfirmDialog(null, "¿Desea continuar?", "Seleccione una opción",
+                    javax.swing.JOptionPane.YES_NO_CANCEL_OPTION, javax.swing.JOptionPane.YES_NO_CANCEL_OPTION);
+            if (input == 0) {
+                DefaultTableModel mdl = (DefaultTableModel) jTable1.getModel();
+                Cotizacion c = (Cotizacion) jTable1.getValueAt(jTable1.getSelectedRow(), 3);
+                //Preguntar si la cotización no está en uso antes de borrar.
+                List<Factura> facturas = Conexion.getInstance().listarFacturasSinFechaADolares();
+                for (Factura f : facturas) {
+                    if ((float) c.getImporte() == f.getCotizacion()) {
+                        javax.swing.JOptionPane.showMessageDialog(null, "No puede borrar una cotización si tiene una o más facturas asignadas.", "Error", javax.swing.JOptionPane.ERROR_MESSAGE);
+                        return;
+                    }
                 }
-            }
-            boolean exito = Conexion.getInstance().deleteBoolean(c);
-            if (exito) {
-                javax.swing.JOptionPane.showMessageDialog(null, "La cotización fue eliminada correctamente.", "Enhorabuena", javax.swing.JOptionPane.INFORMATION_MESSAGE);
-                mdl.removeRow(jTable1.getSelectedRow());
-            } else {
-                javax.swing.JOptionPane.showMessageDialog(null, "Ha ocurrido un problema.", "Error", javax.swing.JOptionPane.ERROR_MESSAGE);
+                boolean exito = Conexion.getInstance().deleteBoolean(c);
+                if (exito) {
+                    javax.swing.JOptionPane.showMessageDialog(null, "La cotización fue eliminada correctamente.", "Enhorabuena", javax.swing.JOptionPane.INFORMATION_MESSAGE);
+                    mdl.removeRow(jTable1.getSelectedRow());
+                } else {
+                    javax.swing.JOptionPane.showMessageDialog(null, "Ha ocurrido un problema.", "Error", javax.swing.JOptionPane.ERROR_MESSAGE);
+                }
             }
         }
     }//GEN-LAST:event_jButton1ActionPerformed
